Move site metadata to graphql
