+++ pages/index.js
@@ -1,10 +1,11 @@
-import {client, gql} from '../utils/graphql-client'
+import {client} from '../utils/graphql/client'
 
 import {Bio} from '../components/bio'
 import Layout from '../components/layout'
 import {SEO} from '../components/seo'
 import {PostListing} from '../components/post-listing'
 import {rhythm} from '../utils/typography'
+import {GET_POST_LISTINGS} from '../utils/graphql/queries'
 
 export default function HomePage({posts, siteMetadata}) {
 	return (
@@ -19,27 +20,7 @@ export default function HomePage({posts, siteMetadata}) {
 }
 
 export async function getStaticProps() {
-	const query = gql`
-		query GetPostListings {
-			posts {
-				slug
-				content {
-					title
-					description
-				}
-				stats {
-					date
-					published
-					timeToRead {
-						text
-					}
-				}
-			}
-			siteMetadata {
-				...Metadata
-			}
-		}
-	`
+	const query = GET_POST_LISTINGS
 	const {data} = await client.query({query})
 	return {props: data}
 }
