+++ utils/graphql/client.js
@@ -1,7 +1,7 @@
 import {ApolloClient, InMemoryCache} from '@apollo/client'
 export {gql} from '@apollo/client'
 import {SchemaLink} from 'apollo-link-schema'
-import {resolvers, typeDefinitions as typeDefs} from '../pages/api/graphql'
+import {resolvers, typeDefinitions as typeDefs} from '../../pages/api/graphql'
 import {makeExecutableSchema} from 'graphql-tools'
 
 const schema = makeExecutableSchema({
